Fix sending action$ from children to parent
I think this was the worst debugging session I ever did. I spent a
weekend trying to understand why the timeline did not send any output
to its parent (here all the api request).

The problem was due to creating the timeline component inside a stream
an not sharing it. As a consequence, every stream that originated from
that stream of component got a new component each time. So if I had a
stream that output the DOM of the component and another one that output
the action$, both of them would access different component, so the
action$ would never be fired